build: fix format string warnings
Related ChromiumOS commit that also fixed these errors:
https://chromium.googlesource.com/chromiumos/platform/update_engine/+/75039d7397f03dff77bdf4e26398049ff88edc4c

Differences from that commit:
 - off_t is signed, so don't cast to unsigned
 - size_t can use %zu, no casting required

+++ delta_diff_generator.cc
@@ -477,17 +477,18 @@ void ReportPayloadUsage(const DeltaArchiveManifest& manifest,
 
   std::sort(objects.begin(), objects.end());
 
-  static const char kFormatString[] = "%6.2f%% %10llu %-10s %s\n";
+  static const char kFormatString[] = "%6.2f%% %10jd %-10s %s\n";
   for (vector<DeltaObject>::const_iterator it = objects.begin();
        it != objects.end(); ++it) {
     const DeltaObject& object = *it;
     fprintf(stderr, kFormatString,
             object.size * 100.0 / total_size,
-            object.size,
+            static_cast<intmax_t>(object.size),
             object.type >= 0 ? kInstallOperationTypes[object.type] : "-",
             object.name.c_str());
   }
-  fprintf(stderr, kFormatString, 100.0, total_size, "", "<total>");
+  fprintf(stderr, kFormatString,
+          100.0, static_cast<intmax_t>(total_size), "", "<total>");
 }
 
 }  // namespace {}
